backend/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow
easyocr
opencv-python-headless
sqlalchemy
python-multipart
